+++ lib/catcher.dart
@@ -1,6 +1,7 @@
 library catcher;
 
 export 'package:catcher/core/catcher.dart';
+export 'package:catcher/core/catcher_screenshot.dart';
 export "package:catcher/handlers/console_handler.dart";
 export "package:catcher/handlers/discord_handler.dart";
 export "package:catcher/handlers/email_auto_handler.dart";