+++ lib/handlers/email_auto_handler.dart
@@ -1,11 +1,11 @@
+import 'package:catcher/handlers/base_email_handler.dart';
 import 'package:catcher/model/platform_type.dart';
 import 'package:catcher/model/report.dart';
-import 'package:catcher/model/report_handler.dart';
 import 'package:logging/logging.dart';
 import 'package:mailer/mailer.dart';
 import 'package:mailer/smtp_server.dart';
 
-class EmailAutoHandler extends ReportHandler {
+class EmailAutoHandler extends BaseEmailHandler {
   final String smtpHost;
   final int smtpPort;
   final String senderEmail;
@@ -13,12 +13,6 @@ class EmailAutoHandler extends ReportHandler {
   final String senderPassword;
   final bool enableSsl;
   final List<String> recipients;
-  final bool enableDeviceParameters;
-  final bool enableApplicationParameters;
-  final bool enableStackTrace;
-  final bool enableCustomParameters;
-  final String? emailTitle;
-  final String? emailHeader;
   final bool sendHtml;
   final bool printLogs;
   final Logger _logger = Logger("EmailAutoHandler");
@@ -31,15 +25,23 @@ class EmailAutoHandler extends ReportHandler {
     this.senderPassword,
     this.recipients, {
     this.enableSsl = false,
-    this.enableDeviceParameters = true,
-    this.enableApplicationParameters = true,
-    this.enableStackTrace = true,
-    this.enableCustomParameters = true,
-    this.emailTitle,
-    this.emailHeader,
     this.sendHtml = true,
     this.printLogs = false,
-  }) : assert(recipients.isNotEmpty, "Recipients can't be null or empty");
+    String? emailTitle,
+    String? emailHeader,
+    bool enableDeviceParameters = true,
+    bool enableApplicationParameters = true,
+    bool enableStackTrace = true,
+    bool enableCustomParameters = true,
+  })  : assert(recipients.isNotEmpty, "Recipients can't be null or empty"),
+        super(
+          enableDeviceParameters,
+          enableApplicationParameters,
+          enableStackTrace,
+          enableCustomParameters,
+          emailTitle,
+          emailHeader,
+        );
 
   @override
   Future<bool> handle(Report error) {
@@ -51,11 +53,15 @@ class EmailAutoHandler extends ReportHandler {
       final message = Message()
         ..from = Address(senderEmail, senderName)
         ..recipients.addAll(recipients)
-        ..subject = _getEmailTitle(report)
-        ..text = _setupRawMessageText(report);
+        ..subject = getEmailTitle(report)
+        ..text = setupRawMessageText(report);
+
+      if (report.screenshot != null) {
+        message.attachments = [FileAttachment(report.screenshot!)];
+      }
 
       if (sendHtml) {
-        message.html = _setupHtmlMessageText(report);
+        message.html = setupHtmlMessageText(report);
       }
       _printLog("Sending email...");
 
@@ -81,94 +87,6 @@ class EmailAutoHandler extends ReportHandler {
         password: senderPassword);
   }
 
-  String? _getEmailTitle(Report report) {
-    if (emailTitle?.isNotEmpty == true) {
-      return emailTitle;
-    } else {
-      return "Error report: >> ${report.error} <<";
-    }
-  }
-
-  String _setupHtmlMessageText(Report report) {
-    final StringBuffer buffer = StringBuffer();
-    if (emailHeader?.isNotEmpty == true) {
-      buffer.write(emailHeader);
-      buffer.write("<hr><br>");
-    }
-
-    buffer.write("<h2>Error:</h2>");
-    buffer.write(report.error.toString());
-    buffer.write("<hr><br>");
-    if (enableStackTrace) {
-      buffer.write("<h2>Stack trace:</h2>");
-      buffer.write(report.stackTrace.toString().replaceAll("\n", "<br>"));
-      buffer.write("<hr><br>");
-    }
-    if (enableDeviceParameters) {
-      buffer.write("<h2>Device parameters:</h2>");
-      for (final entry in report.deviceParameters.entries) {
-        buffer.write("<b>${entry.key}</b>: ${entry.value}<br>");
-      }
-      buffer.write("<hr><br>");
-    }
-    if (enableApplicationParameters) {
-      buffer.write("<h2>Application parameters:</h2>");
-      for (final entry in report.applicationParameters.entries) {
-        buffer.write("<b>${entry.key}</b>: ${entry.value}<br>");
-      }
-      buffer.write("<br><br>");
-    }
-
-    if (enableCustomParameters) {
-      buffer.write("<h2>Custom parameters:</h2>");
-      for (final entry in report.customParameters.entries) {
-        buffer.write("<b>${entry.key}</b>: ${entry.value}<br>");
-      }
-      buffer.write("<br><br>");
-    }
-
-    return buffer.toString();
-  }
-
-  String _setupRawMessageText(Report report) {
-    final StringBuffer buffer = StringBuffer();
-    if (emailHeader?.isNotEmpty == true) {
-      buffer.write(emailHeader);
-      buffer.write("\n\n");
-    }
-
-    buffer.write("Error:\n");
-    buffer.write(report.error.toString());
-    buffer.write("\n\n");
-    if (enableStackTrace) {
-      buffer.write("Stack trace:\n");
-      buffer.write(report.stackTrace.toString());
-      buffer.write("\n\n");
-    }
-    if (enableDeviceParameters) {
-      buffer.write("Device parameters:\n");
-      for (final entry in report.deviceParameters.entries) {
-        buffer.write("${entry.key}: ${entry.value}\n");
-      }
-      buffer.write("\n\n");
-    }
-    if (enableApplicationParameters) {
-      buffer.write("Application parameters:\n");
-      for (final entry in report.applicationParameters.entries) {
-        buffer.write("${entry.key}: ${entry.value}\n");
-      }
-      buffer.write("\n\n");
-    }
-    if (enableCustomParameters) {
-      buffer.write("Custom parameters:\n");
-      for (final entry in report.customParameters.entries) {
-        buffer.write("${entry.key}: ${entry.value}\n");
-      }
-      buffer.write("\n\n");
-    }
-    return buffer.toString();
-  }
-
   void _printLog(String log) {
     if (printLogs) {
       _logger.info(log);
@@ -181,7 +99,7 @@ class EmailAutoHandler extends ReportHandler {
         PlatformType.iOS,
         PlatformType.web,
         PlatformType.linux,
+        PlatformType.macOS,
         PlatformType.windows,
-        PlatformType.macOS
       ];
 }