+++ lib/model/application_profile.dart
@@ -1,1 +1,11 @@
-enum ApplicationProfile { debug, release, profile }
+///Available application run modes.
+enum ApplicationProfile {
+  ///Debug mode
+  debug,
+
+  ///Release mode
+  release,
+
+  ///Profile mode
+  profile,
+}