+++ lib/model/report.dart
@@ -1,3 +1,5 @@
+import 'dart:io';
+
 import 'package:catcher/model/platform_type.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
@@ -27,6 +29,10 @@ class Report {
   /// Type of platform used
   final PlatformType platformType;
 
+  ///Screenshot of screen where error happens. Screenshot won't work everywhere
+  /// (i.e. web platform), so this may be null.
+  final File? screenshot;
+
   /// Creates report instance
   Report(
     this.error,
@@ -37,6 +43,7 @@ class Report {
     this.customParameters,
     this.errorDetails,
     this.platformType,
+    this.screenshot,
   );
 
   /// Creates json from current instance