+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## 0.6.2
+* Fixed Platform.* code which couldn't be used in web.
+* Fixed issues where web couldn't use some of report mode or report handler.
+* Added getInstance method to Catcher.
+* Refactored email handlers. 
+* Added html escape in email handlers.
+* Fixed windows building.
+* Added german translations by Nico Mexis (https://github.com/ThexXTURBOXx)
+* Added screenshot feature. Screenshot will work for all platforms except Web. Screenshot is compatible
+with Http Handler, Email Auto Handler, Email Manual Handler, Discord Handler.
+* Added new field in CatcherOptions: screenshotsPath.
+* Added CatcherScreenshot widget.
+* Added screenshot example.
+
 ## 0.6.1
 * Refactored platforms to fix pub dev platforms score.